aarogyavaani/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
pydantic>=2.5.0
tqdm>=4.66.0
